+++ backend/src/app/core/vector/crud/sentence_embedding.py
@@ -107,11 +107,12 @@ class CRUDSentenceEmbedding(CRUDBase[SentenceObjectIdentifier, SentenceCollectio
             sdoc_id: The SourceDocument ID
         """
         collection = self._get_collection(client=client, project_id=project_id)
-        collection.data.delete_many(
-            where=Filter.by_property(
-                self.collection_class.properties["sdoc_id"].name
-            ).equal(sdoc_id)
-        )
+        if self._tenant_exists(client=client, project_id=project_id):
+            collection.data.delete_many(
+                where=Filter.by_property(
+                    self.collection_class.properties["sdoc_id"].name
+                ).equal(sdoc_id)
+            )
 
 
 crud_sentence_embedding = CRUDSentenceEmbedding(
